working standalone leveldb server

+++ partitioner.go
@@ -8,6 +8,6 @@ func MakePartitioner(topics int) func(string) int32 {
 	return func(tag string) int32 {
 		h := fnv.New32a()
 		h.Write([]byte(tag))
-		return int32(h.Sum32()) % int32(topics)
+		return int32(uint32(h.Sum32()) % uint32(topics))
 	}
 }
